Remove unused files with unused strings
