+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController.swift
@@ -470,46 +470,6 @@ class BrowserViewController: UIViewController {
 #endif
     }
 
-    private func showCrashOptInAlert() {
-        let alert = UIAlertController.crashOptInAlert(
-            sendReportCallback: { _ in
-                // Turn on uploading but don't save opt-in flag to profile because this is a one time send.
-                //configureActiveCrashReporter(true)
-                self.showRestoreTabsAlert()
-            },
-            alwaysSendCallback: { _ in
-                self.profile.prefs.setBool(true, forKey: "crashreports.send.always")
-                //configureActiveCrashReporter(true)
-                self.showRestoreTabsAlert()
-            },
-            dontSendCallback: { _ in
-                // no-op: Do nothing if we don't want to send it
-                self.showRestoreTabsAlert()
-            }
-        )
-        self.presentViewController(alert, animated: true, completion: nil)
-    }
-
-    private func showRestoreTabsAlert() {
-        guard shouldRestoreTabs() else {
-            self.tabManager.addTabAndSelect()
-            return
-        }
-
-        let alert = UIAlertController.restoreTabsAlert(
-            okayCallback: { _ in
-                self.tabManager.restoreTabs()
-                self.updateTabCountUsingTabManager(self.tabManager, animated: false)
-            },
-            noCallback: { _ in
-                self.tabManager.addTabAndSelect()
-                self.updateTabCountUsingTabManager(self.tabManager, animated: false)
-            }
-        )
-
-        self.presentViewController(alert, animated: true, completion: nil)
-    }
-
     private func shouldRestoreTabs() -> Bool {
         guard let tabsToRestore = TabManager.tabsToRestore() else { return false }
         let onlyNoHistoryTabs = !tabsToRestore.every { $0.sessionData?.urls.count > 1 || !AboutUtils.isAboutHomeURL($0.sessionData?.urls.first) }
