+++ Client/Frontend/Settings/AppSettingsTableViewController.swift
@@ -47,8 +47,6 @@ class AppSettingsTableViewController: SettingsTableViewController {
                 titleText: NSLocalizedString("Block Pop-up Windows", comment: "Block pop-up windows setting")),
             BoolSetting(prefs: prefs, prefKey: "saveLogins", defaultValue: true,
                 titleText: NSLocalizedString("Save Logins", comment: "Setting to enable the built-in password manager")),
-            BoolSetting(prefs: prefs, prefKey: AllowThirdPartyKeyboardsKey, defaultValue: false,
-                titleText: NSLocalizedString("Allow Third-Party Keyboards", comment: "Setting to enable third-party keyboards"), statusText: NSLocalizedString("Firefox needs to reopen for this change to take effect.", comment: "Setting value prop to enable third-party keyboards")),
         ]
 
         let accountChinaSyncSetting: [Setting]
@@ -61,16 +59,7 @@ class AppSettingsTableViewController: SettingsTableViewController {
 //                ChinaSyncServiceSetting(settings: self)
             ]
         }
-        // There is nothing to show in the Customize section if we don't include the compact tab layout
-        // setting on iPad. When more options are added that work on both device types, this logic can
-        // be changed.
-        if UIDevice.currentDevice().userInterfaceIdiom == .Phone {
-            generalSettings +=  [
-                BoolSetting(prefs: prefs, prefKey: "CompactTabLayout", defaultValue: true,
-                    titleText: NSLocalizedString("Use Compact Tabs", comment: "Setting to enable compact tabs in the tab overview"))
-            ]
-        }
-
+        
         settings += [
             SettingSection(title: nil, children: [
 //                // Without a Firefox Account:
@@ -83,13 +72,6 @@ class AppSettingsTableViewController: SettingsTableViewController {
         settings += [ SettingSection(title: NSAttributedString(string: NSLocalizedString("General", comment: "General settings section title")), children: generalSettings)]
 
         var privacySettings = [Setting]()
-        if AppConstants.MOZ_LOGIN_MANAGER {
-            privacySettings.append(LoginsSetting(settings: self, delegate: settingsDelegate))
-        }
-
-        if AppConstants.MOZ_AUTHENTICATION_MANAGER {
-            privacySettings.append(TouchIDPasscodeSetting(settings: self))
-        }
 
         privacySettings.append(ClearPrivateDataSetting(settings: self))
 
@@ -102,28 +84,9 @@ class AppSettingsTableViewController: SettingsTableViewController {
         ]
 
         privacySettings += [
-//            BoolSetting(prefs: prefs, prefKey: "crashreports.send.always", defaultValue: false,
-//                titleText: NSLocalizedString("Send Crash Reports", comment: "Setting to enable the sending of crash reports"),
-//                settingDidChange: { configureActiveCrashReporter($0) }),
             PrivacyPolicySetting()
         ]
 
-
-        settings += [
-            SettingSection(title: NSAttributedString(string: privacyTitle), children: privacySettings),
-            SettingSection(title: NSAttributedString(string: NSLocalizedString("Support", comment: "Support section title")), children: [
-                ShowIntroductionSetting(settings: self),
-                SendFeedbackSetting(),
-                OpenSupportPageSetting(delegate: settingsDelegate),
-            ]),
-            SettingSection(title: NSAttributedString(string: NSLocalizedString("About", comment: "About settings section title")), children: [
-                VersionSetting(settings: self),
-                LicenseAndAcknowledgementsSetting(),
-                YourRightsSetting(),
-                ExportBrowserDataSetting(settings: self),
-                DeleteExportedDataSetting(settings: self),
-            ])]
-            
         return settings
     }
